+++ konva.js
@@ -13180,7 +13180,6 @@
         }
 
         node.getStage()._setPointerPosition(evt);
-        node._setDragPosition(evt);
         if (!dd.isDragging) {
           dd.isDragging = true;
           node.fire(
@@ -13192,7 +13191,12 @@
             },
             true
           );
+          // a user can stop dragging inside `dragstart`
+          if (!node.isDragging()) {
+            return;
+          }
         }
+        node._setDragPosition(evt);
 
         // execute ondragmove if defined
         node.fire(